+++ CHANGELOG.md
@@ -5,15 +5,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix handling of `--debug` flag when gunicorn is not present ([#44])
 
 ## [1.4.1] - 2020-05-07
+### Fixed
 - Fix Windows support ([#38])
 
 ## [1.4.0] - 2020-05-06
+### Changed
 - Use gunicorn as a production HTTP server
 
 ## [1.3.0] - 2020-04-13
+### Added
 - Add support for running `python -m functions_framework` ([#31])
+
+### Changed
 - Move `functions_framework.cli.cli` to `functions_framework._cli._cli`
 - Adjust path handling for robots.txt and favicon.ico ([#33])
 
@@ -56,6 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.0.1]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v1.0.1
 [1.0.0]: https://github.com/GoogleCloudPlatform/functions-framework-python/releases/tag/v1.0.0
 
+[#44]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/44
 [#38]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/38
 [#33]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/33
 [#31]: https://github.com/GoogleCloudPlatform/functions-framework-python/pull/31